Add Licence to manifest, fixes powercord detection.
This fixes StyleManager giving error "Invalid Manifest" for "Invalid Licence" as the license string was missing. This was an internal error on my part.

This now also tells StyleManager to find the theme on the root of the cloned folder instead of a folder within itself as there is none.